requirements.txt
certifi==2024.8.30 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
cffi==1.17.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
charset-normalizer==3.3.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
click==8.1.7 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
colorama==0.4.6 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0" and (platform_system == "Windows" or sys_platform == "win32")
dynaconf==3.2.6 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
greenlet==3.1.0 ; python_version < "3.12.dev0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.11.dev0"
idna==3.8 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
jinja2==3.1.4 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
joblib==1.4.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
loguru==0.7.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
lz4==4.3.3 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
markupsafe==2.1.5 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
more-itertools==10.5.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
nltk==3.9.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
numpy==2.1.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pandas==2.2.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pathvalidate==2.5.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pdoc==13.1.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
psycopg2-binary==2.9.9 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pyarrow==17.0.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pycparser==2.22 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pygit2==1.15.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pygments==2.18.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
python-dotenv==1.0.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pytz==2024.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
regex==2024.9.11 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
requests==2.32.3 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
six==1.16.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
sqlalchemy==2.0.34 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
tqdm==4.66.5 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
typing-extensions==4.12.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
tzdata==2024.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
unidecode==1.3.8 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
urllib3==2.2.3 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
win32-setctime==1.1.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0" and sys_platform == "win32"
